% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coal.Theta.Sim.File.G.R
\name{Coal.Theta.Sim.File.G}
\alias{Coal.Theta.Sim.File.G}
\title{Coal.Theta.Sim.File.G: function to simulate a vector (k.vec) of mutation counts given a vector of locus lengths (l.vec), vector of sample numbers (n.vec) with gamma rate variation}
\usage{
Coal.Theta.Sim.File.G(theta, n.vec, l.vec, num.loci, out.file, alpha.param, K)
}
\arguments{
\item{theta}{Population size parameter theta}

\item{num.loci}{Number of loci to simulate}

\item{out.file}{Outfile to write in alleles format}

\item{alpha.param}{Shape parameter of the gamma distribution for among-locus rate variation}

\item{K}{Number of discrete classes to approximate the gamma distribution}

\item{n.list}{List containing the number of samples for each locus}

\item{l.list}{List containing the number of samples for each locus}
}
\description{
This function returns k.vec, l.vec, n.vec from simulating under the coalescent model, with the results written to 'out.file'
}
\examples{


}
\keyword{coalescent}
\keyword{genetics,}
\keyword{models}
\keyword{population}

