% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coal.Theta.Sim.R
\name{Coal.Theta.Sim}
\alias{Coal.Theta.Sim}
\title{Coal.Theta.Sim: function to simulate a vector (k.vec) of mutation counts given a vector of locus lengths (l.vec), vector of sample numbers (n.vec)}
\usage{
Coal.Theta.Sim(theta, n.vec, l.vec, num.loci)
}
\arguments{
\item{theta}{Population size parameter theta}

\item{num.loci}{Number of loci to simulate}

\item{n.list}{List containing the number of samples for each locus}

\item{l.list}{List containing the number of samples for each locus}
}
\description{
This function returns k.vec, l.vec, n.vec from simulating under the coalescent model, with the results written to 'out.file'
}
\examples{
Sim.Results <- Coal.Sim(theta = 0.01, n.vec = c(10,10,10), l.vec = c(100,100,100), num.loci = 3)

Returns:
k.vec = # of mutations for each unique data pattern
n.vec = # of samples for each unique data pattern
l.vec = # of sites for each unique data pattern
c.vec = # of counts of each unique data pattern
}
\keyword{coalescent}
\keyword{genetics,}
\keyword{models}
\keyword{population}

