% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compute.RateAVG.R
\name{Compute.RateAVG.R}
\alias{Compute.RateAVG.R}
\title{Compute.RateAVG: function to compute the average rate parameter within an interval of the gamma distribution with parameter alpha}
\usage{
Compute.RateAVG.R(k.classes, alpha.param)
}
\arguments{
\item{k.classes}{Number of discrete categories in the gamma distribution}

\item{alpha.param}{Alpha parameter of the gamma distribution}
}
\description{
This function returns a list of each step in the MCMC sampling chain
}
\keyword{coalescent}
\keyword{genetics,}
\keyword{models}
\keyword{population}

