% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read.AllelesFile.R
\name{Read.AllelesFile}
\alias{Read.AllelesFile}
\title{Read.AllelesFile: Function to read PyRAD *.alleles files and convert to ThetaMater format}
\usage{
Read.AllelesFile(alleles.file)
}
\arguments{
\item{allele.file}{pyRAD output file (*.alleles)}

\item{log.file}{Log file to write filtering results}
}
\description{
This function returns a list containing (1) the number of mutations for each RAD locus, (2) The length of each RADseq locu, and (3) the number of sequences at each RAD locus
}
\examples{
example.dat <- Read.AllelesFile(alleles.file = '~/Desktop/example.alleles')
}
\keyword{coalescent}
\keyword{genetics,}
\keyword{models}
\keyword{population}

