% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read.DiploidFasta.R
\name{Read.DiploidFasta}
\alias{Read.DiploidFasta}
\title{Read.DiploidFasta: Function to read diploid genome sequence in fasta format}
\usage{
Read.DiploidFasta(genome.fasta.file, num.contigs)
}
\arguments{
\item{genome.fasta.file}{Fasta file comprising multiple contigs with ambiquity codes}
}
\description{
This function returns a list containing (1) the number of mutations for each RAD locus, (2) The length of each RADseq locu, and (3) the number of sequences at each RAD locus
}
\examples{
Read.DiploidFasta(genome.fasta.file = '~/Desktop/example.genome.fasta', num.contigs = 4)
}
\keyword{coalescent}
\keyword{genetics,}
\keyword{models}
\keyword{population}

