% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read.FastaDir.R
\name{Read.FastaDir}
\alias{Read.FastaDir}
\title{Read.FastaDir: Function to read a set of locus-specific fasta files provided in a directory}
\usage{
Read.FastaDir(fasta.dir)
}
\arguments{
\item{fasta.dir}{Directory containing fasta files for each individual locus}
}
\description{
This function returns a list containing (1) the number of mutations for each RAD locus, (2) The length of each locus, and (3) the number of sequences at each locus
}
\examples{
example.dat <- Read.FastaDir(fasta.dir = '~/Desktop/ExampleFastaDir/')
}
\keyword{coalescent}
\keyword{genetics,}
\keyword{models}
\keyword{population}

