% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read.InterleavedFasta.R
\name{Read.InterleavedFasta}
\alias{Read.InterleavedFasta}
\title{Read.InterleavedFasta: Function to read interleaved fasta file (loci are stacked, as with Stacks output files)}
\usage{
Read.InterleavedFasta(fasta.file)
}
\arguments{
\item{fasta.file}{Fasta file (pile up across individual loci - see Stacks output file for example)}
}
\description{
This function returns a list containing (1) the number of mutations for each RAD locus, (2) The length of each RADseq locu, and (3) the number of sequences at each RAD locus
}
\examples{
example.dat <- Read.InterleavedFasta(fasta.file = '~/Desktop/example.fasta')
}
\keyword{coalescent}
\keyword{genetics,}
\keyword{models}
\keyword{population}

