% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read.NexusDir.R
\name{Read.NexusDir}
\alias{Read.NexusDir}
\title{Read.NexusDir: Function to read a set of locus-specific nexus files provided in a directory}
\usage{
Read.NexusDir(nexus.dir)
}
\arguments{
\item{nexus.dir}{Directory containing nexus files for each individual locus}
}
\description{
This function returns a list containing (1) the number of mutations for each RAD locus, (2) The length of each locus, and (3) the number of sequences at each locus
}
\examples{
example.dat <- Read.NexusDir(nexus.dir = '~/Desktop/ExampleNexusDir/')
}
\keyword{coalescent}
\keyword{genetics,}
\keyword{models}
\keyword{population}

