% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ThetaDataLike}
\alias{ThetaDataLike}
\title{ThetaDataLike}
\usage{
ThetaDataLike(theta, kvec, lvec, nvec, cvec)
}
\arguments{
\item{theta}{Population parameter to compute the likelihood of the dataset}

\item{kvec}{Vector of the mutation counts for each locus}

\item{lvec}{Vector of the locus lengths for each locus}

\item{nvec}{Vector of the number of samples for each locus}

\item{cvec}{Vector of the number of pattern counts}
}
\description{
Function to compute the probability of a given dataset
}
\details{
\code{ThetaDataLike} Computes the likelihood of a dataset for a given value of theta
}

