% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThetaMater.M0.R
\name{ThetaMater.M0}
\alias{ThetaMater.M0}
\title{ThetaMater.M0: (Bayesian Model 1) Simple MCMC algorithm to estimate the posteriori distribution of theta from genomic data}
\usage{
ThetaMater.M0(k.vec, l.vec, n.vec, c.vec, ngens)
}
\arguments{
\item{k.vec}{Vector of mutation counts}

\item{l.vec}{Vector of locus lengths}

\item{n.vec}{Vector of sample numbers}

\item{c.vec}{Vector of data pattern counts}

\item{ngens}{Number of generations to run the MCMC}
}
\description{
This function returns a list of each step in the MCMC sampling chain
}
\examples{
sim.results <- Coal.Theta.Sim.File(theta = .9/10, n.vec = c(5,5), l.vec = c(100,100), num.loci = 1000, out.file = '~/Desktop/example.alleles')
example.data <- Read.AllelesFile.Threshold(alleles.file = '~/Desktop/example.alleles', threshold = 1000, log.file = '~/Desktop/log.txt')
Theta.Posterior <- MCMC.Theta.M1.R(k.vec = example.data$k.vec, l.vec = example.data$l.vec, n.vec = example.data$n.vec, c.vec = example.data$c.vec, ngens = 10000)

}
\keyword{coalescent}
\keyword{genetics,}
\keyword{models}
\keyword{population}

