% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThetaMater.M1.R
\name{ThetaMater.M1}
\alias{ThetaMater.M1}
\title{ThetaMater.M1: (Bayesian Model 2) MCMC function that uses the MCMCmetrop1R function from MCMCpack}
\usage{
ThetaMater.M1(k.vec, l.vec, n.vec, c.vec, ngens, burnin, thin, theta.shape,
  theta.scale)
}
\arguments{
\item{k.vec}{Vector of mutation counts}

\item{l.vec}{Vector of locus lengths}

\item{n.vec}{Vector of sample numbers}

\item{c.vec}{Vector of data pattern counts}

\item{ngens}{Number of generations to run the MCMC}

\item{burnin}{Number of generations to discard as burnin}

\item{thin}{Number of generations to thin in the MCMC chain}

\item{theta.shape}{Shape parameter of the gamma distribution for setting the prior on theta}

\item{theta.scale}{Scale parameter of the gamma distribution for setting the prior on theta}
}
\description{
This function returns a list of each step in the MCMC sampling chain
}
\examples{
library(Rcpp)
library(ThetaMater)
library(MCMCpack)
}
\keyword{coalescent}
\keyword{genetics,}
\keyword{models}
\keyword{population}

