% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThetaMater.M2.R
\name{ThetaMater.M2}
\alias{ThetaMater.M2}
\title{ThetaMater.M2: (Bayesian Model 2) MCMC function that uses the MCMCmetrop1R function from MCMCpack with a discrete gamma rate variation model}
\usage{
ThetaMater.M2(k.vec, l.vec, n.vec, c.vec, K, alpha.param, ngens, burnin, thin,
  theta.shape, theta.scale)
}
\arguments{
\item{k.vec}{Vector of mutation counts}

\item{l.vec}{Vector of locus lengths}

\item{n.vec}{Vector of sample numbers}

\item{c.vec}{Vector of data pattern counts}

\item{K}{Number of classese to approximate the gamma distribution}

\item{alpha.param}{Shape of the gamma distribution for describing the amount of among-locus rate variation}

\item{ngens}{Number of generations to run the MCMC}

\item{burnin}{Number of generations to discard from MCMC chain}

\item{thin}{Number of generations to thin in the MCMC chain}

\item{theta.shape}{Shape parameter of the gamma distribution for setting the prior on theta}

\item{theta.scale}{Scale parameter of the gamma distribution for setting the prior on theta}
}
\description{
This function returns a list of each step in the MCMC sampling chain
}
\examples{
library(Rcpp)
library(ThetaMater)
library(MCMCpack)

}
\keyword{coalescent}
\keyword{genetics,}
\keyword{models}
\keyword{population}

