% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThetaMater.M3.R
\name{ThetaMater.M3}
\alias{ThetaMater.M3}
\title{ThetaMater.M3: (Bayesian Model 2) MCMC function that uses the MCMCmetrop1R function from MCMCpack with a discrete gamma rate variation model}
\usage{
ThetaMater.M3(k.vec, l.vec, n.vec, c.vec, K, ngens, burnin, thin, theta.shape,
  theta.scale, alpha.shape, alpha.scale)
}
\arguments{
\item{k.vec}{Vector of mutation counts}

\item{l.vec}{Vector of locus lengths}

\item{n.vec}{Vector of sample numbers}

\item{c.vec}{Vector of data pattern counts}

\item{ngens}{Number of generations to run the MCMC}

\item{burnin}{Number of generations to discard from MCMC chain}

\item{thin}{Number of generations to thin in the MCMC chain}

\item{theta.shape}{Shape parameter of the gamma distribution for setting the prior on theta}

\item{theta.scale}{Scale parameter of the gamma distribution for setting the prior on theta}

\item{alpha.shape}{Shape parameter of the gamma distribution for setting the prior on alpha}

\item{alpha.shape}{Scale parameter of the gamma distribution for setting the prior on alpha}
}
\description{
This function returns a list of each step in the MCMC sampling chain
}
\examples{
library(Rcpp)
library(ThetaMater)
library(MCMCpack)
sim.results <- Coal.Theta.Sim.File.G(theta = 0.001, n.vec = c(10,10), l.vec = c(2000,2000), num.loci = 1000, out.file = '~/Desktop/example.alleles', alpha.param = 0.5)
example.data <- Read.AllelesFile.NoThreshold(alleles.file = '~/Desktop/example.alleles', log.file = '~/Desktop/log.txt')
Theta.Posterior.G <- MCMC.Theta.M3(k.vec = example.data$k.vec, l.vec = example.data$l.vec, n.vec = example.data$n.vec, c.vec = example.data$c.vec, 
ngens = 10000, thin = 10, burnin = 10000, K = 4, alpha.param = 0.5)




}
\keyword{coalescent}
\keyword{genetics,}
\keyword{models}
\keyword{population}

