% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThetaMater.PPS.R
\name{ThetaMater.PPS}
\alias{ThetaMater.PPS}
\title{ThetaMater.PPS: function to simulate posterior predictive distributions of mutation counts using the posterior distributions estimated using ThetaMater}
\usage{
ThetaMater.PPS(theta.MCMC, l.vec, n.vec)
}
\arguments{
\item{theta.MCMC}{MCMC chain sampling of theta parameter provided by ThetaMater}

\item{l.vec}{Vector of locus lengths}

\item{n.vec}{Vector of sample numbers}
}
\description{
This function returns a vector of mutation counts for every simulated locus (the posterior predictive distribution)
}
\examples{
sim.results <- Coal.Theta.Sim.File(theta = .9/10, n.vec = c(5,5), l.vec = c(100,100), num.loci = 1000, out.file = '~/Desktop/example.alleles')
example.data <- Read.AllelesFile.Threshold(alleles.file = '~/Desktop/example.alleles', threshold = 1000, log.file = '~/Desktop/log.txt')
Theta.Posterior <- ThetaMater.M3(k.vec = example.data$k.vec, l.vec = example.data$l.vec, n.vec = example.data$n.vec, c.vec = example.data$c.vec, ngens = 10000)

}
\keyword{coalescent}
\keyword{genetics,}
\keyword{models}
\keyword{population}

